requirements.txt
torch
opencv-python
torchvision
gdown
matplotlib
numpy
scipy
tqdm
Pillow
scikit_learn
tensorboardX
grad-cam
ttach
colorama
tensorboard
easydict
mxnet
onnx
sklearn
